% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_search_api.R
\name{call_api}
\alias{call_api}
\title{Call OpenAI's GPT API}
\usage{
call_api(
  model,
  temperature,
  max_tokens,
  messages,
  num_retries = 3,
  pause_base = 1
)
}
\arguments{
\item{model}{Character. The identifier of the language model to use.}

\item{temperature}{Numeric. Controls randomness in the model's output. Range: 0.0 to 1.0.}

\item{max_tokens}{Integer. The maximum number of tokens to be generated by the model.}

\item{messages}{List. Messages to be passed to the model. Each message has two properties: 'role' and 'content'.}

\item{num_retries}{Integer. Number of retry attempts in case of a network error. Default: 3.}

\item{pause_base}{Numeric. The base of the exponential backoff for retries. Default: 1.}
}
\value{
A cleaned message generated by the model.
}
\description{
This function calls OpenAI's GPT API to generate a message using a given language model.
It handles retries and error checking, and processes the response from the API for easier use.
}
\note{
This function requires a valid OpenAI API key, which should be stored in the \code{api_key} variable.
}
\examples{
\dontrun{
  # Define a list of messages to be sent to the model
  messages <- list(
    list('role' = 'system', 'content' = 'You are a helpful assistant.'),
    list('role' = 'user', 'content' = 'Who won the world series in 2020?')
  )

  # Call the function with appropriate parameters
  call_api('text-davinci-002', 0.5, 100, messages)
}

}
