% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_api.R
\name{count_tokens}
\alias{count_tokens}
\title{Estimate token count of a text string}
\usage{
count_tokens(text)
}
\arguments{
\item{text}{A character string for which to estimate the token count.}
}
\value{
An integer giving the estimated number of tokens in the input text.
}
\description{
This function provides a rough estimate of the number of tokens in a text string,
based on splitting the text into words, punctuation and whitespaces.
Note that this is an approximation and may not match the token count used
by specific language models.
}
\examples{
count_tokens("Hello, world!")
count_tokens("This is a longer sentence with more tokens.")

}
