% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalle_api.R
\name{dalle_call}
\alias{dalle_call}
\title{API Call to OpenAI}
\usage{
dalle_call(endpoint, method, headers, body = NULL, query = NULL)
}
\arguments{
\item{endpoint}{A string specifying the API endpoint. This should not include the base URL, as it is automatically prepended.}

\item{method}{A string specifying the HTTP method for the API call, either "GET" or "POST".}

\item{headers}{A named list of HTTP headers to include in the request.}

\item{body}{An optional argument. If the HTTP method is "POST", this should be a list that will be included in the request body.}

\item{query}{An optional argument. If provided, it should be a list to be included as query parameters in the URL.}
}
\value{
The function returns the content of the API response as a list. If the server response is not a JSON, it throws an error.
}
\description{
This function is a common handler for making API calls to OpenAI. It was primarily made to be used with the other functions in this script.
}
\examples{
\dontrun{
# Example of a GET request:
headers <- c("Authorization" = paste0("Bearer ", api_key))
response <- dalle_call("engines", "GET", headers)

# Example of a POST request:
headers <- c("Authorization" = paste0("Bearer ", api_key),
             "Content-Type" = "application/json")
body <- list(prompt = "Translate the following English text to French: '{}'",
             max_tokens = 60)
response <- dalle_call("engines/davinci-codex/completions", "POST", headers, body)
}
}
