% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalle_api.R
\name{dalle_variation}
\alias{dalle_variation}
\title{Generate Variations}
\usage{
dalle_variation(
  image_url,
  save_path = "C:/your/path",
  scale_parameter = "50\%",
  brightness_parameter = 80,
  radius_parameter = 0,
  sigma_parameter = 1
)
}
\arguments{
\item{image_url}{The URL of the image to be processed.}

\item{save_path}{The path where the resulting image will be saved.}

\item{scale_parameter}{The scale parameter for image scaling. Default is "50\%".}

\item{brightness_parameter}{The brightness parameter for modulating brightness. Default is 80.}

\item{radius_parameter}{The radius parameter for applying blur. Default is 0.}

\item{sigma_parameter}{The sigma parameter for applying blur. Default is 1.}
}
\description{
This function generates variations of an image by applying transformations such as scaling, modulating brightness, and applying blur. The resulting image is saved to disk.
}
\examples{
\dontrun{
# Generate variations of an image from a URL
dalle_variation("https://example.com/image.jpg")

# Generate variation from what you queried using the dalle_api() function
dalle_variation(result$data$url[1])
}
}
