% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_search_api.R
\name{gpt}
\alias{gpt}
\title{GPT-based Text Generation}
\usage{
gpt(
  prompt,
  model = "gpt-3.5-turbo",
  temperature = 0.5,
  max_tokens = 50,
  system_messages = NULL
)
}
\arguments{
\item{prompt}{Character. The user message to be processed by the language model.}

\item{model}{Character. The identifier of the language model to use. Default: "gpt-3.5-turbo".}

\item{temperature}{Numeric. Controls randomness in the model's output. Range: 0.0 to 1.0. Default: 0.5.}

\item{max_tokens}{Integer. The maximum number of tokens to be generated by the model. Default: 50.}

\item{system_messages}{List. Optional. System messages to include in the conversation. Default: NULL.}
}
\value{
A message generated by the model based on the given prompt.
}
\description{
This function uses a specified GPT model to generate a text based on a given prompt.
It can optionally include system messages in the conversation. It's built on top of the 'call_api' function.
}
\note{
This function requires a valid OpenAI API key, which should be stored in the \code{api_key} variable.
}
\examples{
\dontrun{
  # Use the GPT function with a user message
  gpt('Tell me a joke', model = 'gpt-3.5-turbo', temperature = 0.6, max_tokens = 40)

  # Use the GPT function with a user message and a system message
  gpt('Tell me a joke', model = 'gpt-3.5-turbo', temperature = 0.6, max_tokens = 40,
      system_messages = 'You are a humoristic AI')
}

}
\seealso{
\code{\link{call_api}}
}
