% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_api.R
\name{gpt_api}
\alias{gpt_api}
\title{This function sends a prompt to the OpenAI API and returns the generated text.
It handles API communication, error checking, and basic formatting of the response.}
\usage{
gpt_api(
  prompt,
  model = "gpt-3.5-turbo",
  temperature = 0.5,
  max_tokens = 50,
  system_message = NULL,
  num_retries = 3,
  pause_base = 1,
  presence_penalty = 0,
  frequency_penalty = 0,
  api_key
)
}
\arguments{
\item{prompt}{The prompt to send to the API.}

\item{model}{The model to use for generating the response. Default is "gpt-3.5-turbo".
This can be adjusted according to the available models in the OpenAI API (such as "gpt-4")}

\item{temperature}{The temperature to use for generating the response. Default is 0.5.}

\item{max_tokens}{The maximum number of tokens in the response. Default is 50.}

\item{system_message}{Optional initial system message to set the behavior of the AI.
Default is NULL.}

\item{num_retries}{Number of times to retry the request in case of failure. Default is 3.}

\item{pause_base}{The number of seconds to wait between retries. Default is 1.}

\item{presence_penalty}{The penalty for new tokens based on their presence in the input. Default is 0.0.}

\item{frequency_penalty}{The penalty for new tokens based on their frequency in the input. Default is 0.0.}

\item{api_key}{Character string for the API key assigned to the user by OpenAI}
}
\value{
A character string containing the generated text.
}
\description{
Register for an API secert here: https://platform.openai.com/account/api-keys
}
\examples{
\dontrun{
api_key <- Sys.getenv('OPENAI_API_KEY')
gpt_api("Write a poem about my cat Sam", api_key = api_key)
gpt_api("Tell me a joke.", model = "gpt-3.5-turbo", api_key = api_key)
gpt_api("Write a poem about my cat Sam", model = "gpt-4", api_key = api_key)
}
}
