% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_search_api.R
\name{gpt_search}
\alias{gpt_search}
\title{Grid Search Over GPT Models}
\usage{
gpt_search(
  prompt,
  models = c("gpt-3.5-turbo", "gpt-4"),
  temperatures = seq(0.2, 1, by = 0.2),
  max_tokens_range = c(20, 100),
  system_messages = list(NULL),
  batch_size = 100,
  save_to_disk = FALSE,
  file_path = "YOUR_PATH_HERE"
)
}
\arguments{
\item{prompt}{Character. The user message to be processed by the language model.}

\item{models}{Character vector. The identifiers of the language models to use. Default: c("gpt-3.5-turbo", "gpt-4").}

\item{temperatures}{Numeric vector. Controls randomness in the model's output. Default: seq(0.2, 1, by = 0.2).}

\item{max_tokens_range}{Integer vector of length 2. The range of maximum tokens to be generated by the models. Default: c(20, 100).}

\item{system_messages}{List of character vectors. System messages to include in the conversation. Default: list(NULL).}

\item{batch_size}{Integer. The number of results to accumulate before saving a batch to disk. Default: 100.}

\item{save_to_disk}{Logical. Whether to save results to disk. Default: FALSE.}

\item{file_path}{Character. The directory where batch files will be saved, if save_to_disk is TRUE. Default: "YOUR_PATH_HERE".}
}
\value{
A data frame of results with columns: model, temperature, max_tokens, system_message, and result.
If save_to_disk is TRUE, it saves the results in batches as .rds files and returns NULL.
}
\description{
This function performs a grid search over the specified GPT models and parameters, using a given prompt.
Results can optionally be saved to disk in batches.
}
\note{
This function requires a valid OpenAI API key, which should be stored in the \code{api_key} variable.
}
\examples{
\dontrun{
  # Set up the system messages
  system_messages <- list("You are a funny comedian", "You are Dane Cook")

  # Perform a grid search
  results <- gpt_search(
    prompt = "Tell me a joke",
    models = c("gpt-3.5-turbo", "gpt-4"),
    temperatures = c(0.2, 0.5, 0.8),
    max_tokens_range = c(20, 30),
    system_messages = system_messages,
    batch_size = 100,
    save_to_disk = TRUE,
    file_path = "your_path_here"
  )
}

}
\seealso{
\code{\link{gpt}}
}
