% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_search_api.R
\name{import_rds_files}
\alias{import_rds_files}
\title{Import RDS Files}
\usage{
import_rds_files(path)
}
\arguments{
\item{path}{The path to the directory containing the RDS files.}
}
\value{
A list of data frames containing the imported data from RDS files.
}
\description{
Set the working directory to the specified path and import RDS files.
}
\examples{
\dontrun{
# Set working directory to your path
setwd("your_path_here")

# Import one batch
readRDS("your_path/batch_1.rds") -> save

# Import all RDS
dir_path <- "path_to_your_directory"
data_list <- import_rds_files(dir_path)
}

}
