% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_read.R
\name{search_text}
\alias{search_text}
\title{Extract key words and passages}
\usage{
search_text(text, keywords, window_size = 500)
}
\arguments{
\item{text}{The input text from which to extract passages.}

\item{keywords}{A character vector of keywords to search for in the text.}

\item{window_size}{An integer specifying the size of the window (in characters)
around each keyword occurrence. The default value is 500.}
}
\value{
A list of passages (windows) containing the occurrences of the keywords.
Each element in the list represents a passage.
}
\description{
This function extracts passages from a given text that contain specified keywords.
It searches for each keyword in the text and returns a list of passages (windows)
that surround the occurrences of the keywords.
}
\examples{
text <- "This is a sample text containing some keywords. The keywords are important for analysis."
keywords <- c("sample", "important")
search_text(text, keywords)

}
