% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper_api.R
\name{whisper_transcribe}
\alias{whisper_transcribe}
\title{Transcribe Audio Using OpenAI's API}
\usage{
whisper_transcribe(
  file_path,
  api_key,
  model = "whisper-1",
  response_format = "json",
  temperature = 0,
  language = NULL,
  prompt = NULL
)
}
\arguments{
\item{file_path}{Character. The file path to the audio file to be transcribed.}

\item{api_key}{Character. The OpenAI API key.}

\item{model}{Character. The name of the OpenAI model to use for transcription. Default is 'whisper-1'.}

\item{response_format}{Character. The format of the response from the API. Default is 'json'.}

\item{temperature}{Numeric. The temperature value to use in the transcription model. Default is 0.}

\item{language}{Character. The language of the audio file. Default is NULL.}

\item{prompt}{Character. The prompt to provide to the transcription model. Default is NULL.}
}
\value{
A list containing the parsed response from the API in the specified response format.
}
\description{
This function uses the OpenAI API to transcribe an audio file.
}
\note{
The \code{api_key} is sensitive information, do not share it or publish it. Make sure to store it securely.
The \code{file_path} should be a valid path to a .wav file.
}
\examples{
\dontrun{
api_key <- "your-api-key"
file_path <- "C:/path_to_your_file/audiofile.wav"
transcription <- whisper_transcribe(file_path, api_key)
print(transcription)

# You may download an example audio file to transcribe here
# https://www.voiptroubleshooter.com/open_speech/american/OSR_us_000_0061_8k.wav
}

}
\seealso{
\url{https://openai.com/research/whisper/} for more information on the Whisper ASR API.
}
