% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper_api.R
\name{whisper_translate}
\alias{whisper_translate}
\title{Translate Audio Using OpenAI's API}
\usage{
whisper_translate(
  file_path,
  api_key,
  model = "whisper-1",
  response_format = "json",
  temperature = 0,
  prompt = NULL
)
}
\arguments{
\item{file_path}{Character. The file path to the audio file to be translated.}

\item{api_key}{Character. The OpenAI API key.}

\item{model}{Character. The name of the OpenAI model to use for translation. Default is 'whisper-1'.}

\item{response_format}{Character. The format of the response from the API. Default is 'json'.}

\item{temperature}{Numeric. The temperature value to use in the translation model. Default is 0.}

\item{prompt}{Character. The prompt to provide to the translation model. Default is NULL.}
}
\value{
A list containing the parsed response from the API in the specified response format.
}
\description{
This function uses the OpenAI API to translate an audio file.
}
\note{
The \code{api_key} is sensitive information, do not share it or publish it. Make sure to store it securely.
The \code{file_path} should be a valid path to a .wav file.
}
\examples{
\dontrun{
api_key <- "your-api-key"
file_path <- "C:/path_to_your_file/audiofile.wav"
translation <- whisper_translate(file_path, api_key)
print(translation)
}

}
\seealso{
\url{https://openai.com/research/whisper/} for more information on the Whisper ASR API.
}
